Fixed role-based navigation and login flow
